Move tests and docs directories outside of source directory
These directories were previously included in the wheel package. Now,
they are not. This had resulted in the files being unnecessarily
downloaded and installed to virtual environments when using pip.
However, these files go unused production installations. These files are
only necessary for the development of importlib_metadata itself, so they
now live at the root of the project.

This was more noticeable when pip dependencies are vendored or bundled,
as it adds unnecessary bloat to these bundles.

The directories are still included in the sdist, so packagers and
library users can run the tests locally if they choose, they just wont
be installed to the site-packages directory.