Distribution.files: Only return files that actually exist
Add an extra filter on the paths returned from Distribution.files, to
prevent paths that don't exist on the filesystem from being returned.

This attempts to solve the issue of .files returning incorrect
information based on the inaccuracies of SOURCES.txt.

As the code currently is organized, it is more complicated to write
this such that it only applies to the information read from SOURCES.txt
specifically, hence we apply it to _all_ of .files instead.

This fixes #115, also in the case where there is no installed-files.txt
file available.

[1]: https://pip.pypa.io/en/stable/news/#v0-3
[2]: https://setuptools.pypa.io/en/latest/deprecated/python_eggs.html#sources-txt-source-files-manifest