Add tests for egg-info package with no installed modules
This corresponds to the qiskit[1] meta-package which:

  - does not contain any (runtime) Python code itself, but serves as a
    mechanism to install its transitive dependencies (which populate the
    qiskit package namespace).

  - is distributed as a source archive.

  - includes a top_level.txt which is empty (contains a single newline),
    arguably correct given that it does not directly install any
    importable packages/modules.

  - when installed as an egg, provides a SOURCES.txt which is incorrect
    from a runtime POV: it references 3 .py files, a setup.py and two
    files under test/, none of which are actually installed.

  - when installed (as an egg) by pip, provides an installed-files.txt
    file which is _more_ accurate than SOURCES.txt, since it reflects
    the files that are actually available after installation.

importlib_metadata reports incorrect .files for this package, because
we end up using SOURCES.txt. It is better to use installed-files.txt
when it is available.

Furthermore, as a result of this, packages_distributions() also
incorrectly reports that this packages provides imports names that do
not actually exist ("setup" and "test", in qiskit's case).

This commit adds EggInfoPkgPipInstalledNoModules, a test project that
mimics the egg installation of qiskit, and adds it to existing test
cases, as well as adding a new test cases specifically for verifying
packages_distributions() with egg-info packages. The following tests
fail in this commit, but will be fixed in the next commit:

  - PackagesDistributionsTest.test_packages_distributions_on_eggs
  - APITests.test_files_egg_info

See the python/importlib_metadata#115 issue for more details.

[1]: qiskit is found at https://pypi.org/project/qiskit/0.41.1/#files