Distribution.files: Prefer *.egg-info/installed-files.txt to SOURCES.txt
When listing the files in a *.egg-info distribution, prefer using
*.egg-info/installed-files.txt instead of *.egg-info/SOURCES.txt.

installed-files.txt is written by pip[1] when installing a package,
whereas the SOURCES.txt is written by setuptools when creating a
source archive[2].

installed-files.txt is only present when the package has been installed
by pip, so we cannot depend on it always being available. However, when
it _is_ available, it is an accurate record of what files are installed.
SOURCES.txt, on the other hand, is always avaiable, but is not always
accurate: Since it is generated from the source archive, it will often
include files (like 'setup.py') that are no longer available after the
package has been installed.

Fixes #115 for the cases where a installed-files.txt file is available.

[1]: https://pip.pypa.io/en/stable/news/#v0-3
[2]: https://setuptools.pypa.io/en/latest/deprecated/python_eggs.html#sources-txt-source-files-manifest