Add tests for egg-info package with .files from inaccurate SOURCES.txt
As established in previous commits, the SOURCES.txt file is not always
an accurate source of files that are present after a package has been
installed.

One situation where this inaccuracy is problematic is when top_level.txt
is also missing, and packages_distributions() is forced to infer the
provided import names based on Distribution.files. In this situation we
end up with incorrect mappings between import packages and distribution
packages, including import packages that clearly do not exist at all.

For example, a SOURCES.txt that lists setup.py (which is used _when_
installing, but is not available after installation), will see that
setup.py returned from .files, which then will cause
packages_distributions() to claim a mapping from the non-existent
'setup' import name to this distribution.

This commit adds EggInfoPkgSourcesFallback which demostrates such a
scenario, and adds this new class to a couple of relevant tests.
A couple of these tests are currently failing, to demonstrate the
issue at hand. These test failures will be fixed in the next commit.

See the python/importlib_metadata#115 issue for more details.